model/requirements.txt
flask
pillow
numpy
tensorflow
google-cloud-storage
gunicorn
keras
flask-cors
